requirements.txt
python-gitlab>=3.15.0
openai>=1.3.7
pydantic>=2.5.2
python-dotenv>=1.0.0
